fix non-hashable Tagged type
`Tagged` is now an interface, `Key` a new type that satisfies that
interface. It seems that values of type `interface{}` (or types
containing fields/values of that type) can't be used as keys in maps
because Go doesn't know how to hash them. We could either use a hash
map, or temporarily fix it this way.